+++ generators/app/templates/infrastructure/src/middleware/logger/loggingMiddleware.js
@@ -1,0 +1,8 @@
+const { logger } = require("../../startup");
+
+const loggingMiddleware = async (ctx, next) => {
+  ctx.logger = logger;
+  await next();
+};
+
+module.exports = loggingMiddleware;
